Moved status route into health-check directory.
